HW 3 fixes
