+++ server.js
@@ -1,17 +1,15 @@
 const app = require('./app')
 
 const mongoose = require("mongoose");
+require("dotenv").config();
 
 mongoose.set('strictQuery', true);
 
-const DB_HOST = "mongodb+srv://PeterZabila:5g8lJFbLU8yh6OUG@cluster1.pqow9jq.mongodb.net/"
-
-
-const {PORT = 3000} = process.env
+const {DB_HOST, PORT = 3000} = process.env
 
 mongoose.connect(DB_HOST)
-.then(() => {console.log("Database connection successful");
-app.listen(PORT)}
+    .then(() => {console.log("Database connection successful");
+    app.listen(PORT)}
 )
 .catch((error) => { 
     console.log(error.message)
